Collision detection fixes
This branch begins a set of tweaks to simplify the collision detection
on the Hubo model.  As it stands, the model runs slowly and often
triggers "hash table bucket overflow" warnings in openrave, though this
may be a bug with the code itself.  Possible ways to eliminate this:

1) Try a different collision checker, either PQP or bullet physics (the
ode physics looks terrible in the physics demo too...)

2) Simplify the model so that the arms and leg collision models are
largely made up of blocks and cylinders, or simplified convex hull
trimeshes.