Lower Body collision primitives (draft version)
Completed import of first draft of Primitive collision models. The
exported data for all of the primitives should be stored in the
LegPrimitive.xlsx file, and updated as the model is improved. The
collision data should be sufficient for testing purposes, though small
improvements still need to be applied.

Long term TODO:
1) Improve torso and hip models (low priority)
2) Replace head collision with sphere